+++ view/frontend/web/js/view/payment/method-renderer.js
@@ -73,7 +73,7 @@ define(
             {type: 'multisafepay_inghomepay', component: baseRenderer},
             {type: 'multisafepay_kbc', component: baseRenderer},
             {type: 'multisafepay_klarna', component: baseRenderer},
-            {type: 'multisafepay_maestro', component: baseRenderer},
+            {type: 'multisafepay_maestro', component: gatewayPath + 'creditcard'},
             {type: 'multisafepay_mastercard', component: gatewayPath + 'creditcard'},
             {type: 'multisafepay', component: baseRenderer},
             {type: 'multisafepay_payafter', component: gatewayPath + 'payafter'},