+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Added MultiSafepay Credit Card component support for credit card payment methods.
+
 ### Changed
 - Moved setting pending_payment status from Redirect controller to Gateway Request Builder
 - Changed the notification method from 'GET' to 'POST'